socket communication debug

+++ include/poledata.h
@@ -5,11 +5,11 @@
 typedef struct
 {
   // position of the upper blob in m
-  float pos_upper[DIM];
+  double pos_upper[DIM];
   // position of the lower blob in m
-  float pos_lower[DIM];
+  double pos_lower[DIM];
   // anlew of the pole in rad
-  float pole_angle;
+  double pole_angle;
   // angular velocity of pole [rad/s]
-  float pole_angular_velocity;
+  double pole_angular_velocity;
 } PoleData;
